+++ packages/vue/src/components/combobox/combobox.ts
@@ -44,6 +44,7 @@ export {
 } from './combobox-item-text.vue'
 export { default as Item, type ComboboxItemProps as ItemProps } from './combobox-item.vue'
 export { default as Label, type ComboboxLabelProps as LabelProps } from './combobox-label.vue'
+export { default as List, type ComboboxListProps as ListProps } from './combobox-list.vue'
 export {
   default as Positioner,
   type ComboboxPositionerProps as PositionerProps,